Changed var type

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
         BodyMassIndex service = new BodyMassIndex();
-        int height = 173; // Рост в см
+        short height = 173; // Рост в см
         double weight = 60; // Вес в кг
         double bodyMassIndex = service.calculate(height, weight);
         String formattedBMI = String.format("%.2f", bodyMassIndex);
